Add hierarchical configurations for service, dependency, and chain settings. Added basic integration with repo to retrieve user records.

+++ v3/internal/dataaccess/repo/val_dashboard_repo_i.go
@@ -3,8 +3,8 @@ package dataaccess
 import "context"
 
 type ValidatorDashboard struct {
-	DashboardId int
-	UserId      int
+	DashboardId uint64
+	UserId      uint64
 	IsPublic    bool
 }
 
@@ -12,16 +12,16 @@ type ValidatorDashboardRepository interface {
 	/**
 	 * Creates an empty Validator Dashboard
 	 */
-	CreateValidatorDashboard(ctx context.Context, userId int) (*ValidatorDashboard, error)
+	CreateValidatorDashboard(ctx context.Context, userId uint64) (*ValidatorDashboard, error)
 
 	/**
 	 * Returns basic info of a dashboard for a particular user.
 	 * userId must be provided
 	 * If no dashboardId is provided, return all dashboards for that user
 	 */
-	GetValidatorDashboardsByUserId(ctx context.Context, userId int) (*[]ValidatorDashboard, error)
+	GetValidatorDashboardsByUserId(ctx context.Context, userId uint64) (*[]ValidatorDashboard, error)
 
-	GetValidatorDashboardByDashboardId(ctx context.Context, dashboardId int) (*ValidatorDashboard, error)
+	GetValidatorDashboardByDashboardId(ctx context.Context, dashboardId uint64) (*ValidatorDashboard, error)
 
 	/**
 	 * Modifies the attributes of a dashboard.
@@ -29,8 +29,8 @@ type ValidatorDashboardRepository interface {
 	 * If an attribute is included but modification of that is not possible, an error should be returned
 	 *    and other attribute modifications should not take place.
 	 */
-	ModifyValidatorDashboard(ctx context.Context, dashboardId int, isPublic bool) (*ValidatorDashboard, error)
+	ModifyValidatorDashboard(ctx context.Context, dashboardId uint64, isPublic bool) (*ValidatorDashboard, error)
 
 	// Returns nothing on success, or error if successfully deleted. Idempotent, if deleted when it DNE, no error should be returned.
-	DeleteValidatorDashboard(ctx context.Context, dashboardId int) error
+	DeleteValidatorDashboard(ctx context.Context, dashboardId uint64) error
 }
